SearchEngine/src/requirements.txt
langchain==0.0.166
PyPDF2==3.0.1
python-dotenv==1.0.0
altair<5
openapi==1.1.0
streamlit==1.22
faiss-cpu==1.7.4
openai==0.27.6
tiktoken==0.4.0
docx2txt==0.8
flask==2.3.2